Original change: https://googleplex-android-review.googlesource.com/c/platform/system/chre/+/17606122

+++ util/include/chre/util/system/atomic_spsc_queue.h
@@ -73,7 +73,7 @@ class AtomicSpscQueue : public NonCopyable {
   ~AtomicSpscQueue() {
     size_t sz = size();
     auto c = consumer();
-    for (int i = 0; i < sz; i++) {
+    for (size_t i = 0; i < sz; i++) {
       c.pop();
     }
   }
